+++ src/misc/service.h
@@ -88,10 +88,34 @@ static void populate_plist(char *yabai_plist, int size)
     snprintf(yabai_plist, size, _YABAI_PLIST, exe_path, path_env, user, user);
 }
 
+static inline void ensure_directory_exists(char *yabai_plist_path)
+{
+    //
+    // NOTE(koekeishiya): Temporarily remove filename.
+    // We know the filepath will contain a slash, as
+    // it is controlled by us, so don't bother checking
+    // the result..
+    //
+
+    char *last_slash = strrchr(yabai_plist_path, '/');
+    *last_slash = '\0';
+
+    if (!directory_exists(yabai_plist_path)) {
+        mkdir(yabai_plist_path, 0755);
+    }
+
+    //
+    // NOTE(koekeishiya): Restore original filename.
+    //
+
+    *last_slash = '/';
+}
+
 static int service_install_internal(char *yabai_plist_path)
 {
     char yabai_plist[4096];
     populate_plist(yabai_plist, sizeof(yabai_plist));
+    ensure_directory_exists(yabai_plist_path);
 
     FILE *handle = fopen(yabai_plist_path, "w");
     if (!handle) return 1;